Fixed a bug which prevented games from being run over 16,000(ish) times by changing the stored datatype from 16bit to 32bit

+++ Unit 2 - Darts/Config.h
@@ -11,7 +11,7 @@ public:
 	uint16_t StartingScore = 501;
 	//uint16_t DartsPerRound = 3;
 	uint16_t MaxRounds = 30;
-	uint16_t NumberOfGames = 1;
+	uint32_t NumberOfGames = 1;
 
 	uint8_t JoeAcc = 71;
 	uint8_t SidAcc = 73;
